Default Changelist

+++ booking/src/main/java/com/pingvin/autoservice/controller/MainController.java
@@ -118,8 +118,10 @@ public class MainController {
         if (utilForm.getTextField().equals("ПОДТВЕРЖДАЮ")) {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             Order order = orderDAO.findOrderByIdOrder(utilForm.getIntField());
-            masterDAO.checkIfMasterIsFree(order.getIdOrder(), order.getMaster().getMaster());
-            orderDAO.removeOrder(order);
+            if(order != null){
+                masterDAO.checkIfMasterIsFree(order.getIdOrder(), order.getMaster().getMaster());
+                orderDAO.removeOrder(order);
+            }
         }
         return "index";
     }
